requirements.txt
joblib==1.3.1
numpy==1.25.0
pandas==2.0.3
omegaconf==2.3.0
scikit-learn==1.3.0
nltk==3.8.1
unidecode==1.3.6
hydra-core==1.3.2
